third_party/blink/web_tests/fast/dom/shadow/suppress-mutation-events-in-shadow-characterdata-expected.txt

CONSOLE WARNING: line 11: Element.createShadowRoot is deprecated and will be removed in M80, around February 2020. Please use Element.attachShadow instead. See https://www.chromestatus.com/features/4507242028072960 and https://developers.google.com/web/updates/2019/07/web-components-time-to-upgrade for more details.
Ensures that DOMCharacterDataModified isn't fired inside shadows.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS fired is false
PASS div.innerHTML is 'Hello, World!'
PASS successfullyParsed is true

TEST COMPLETE
